requirements.txt
python-dotenv==1.0.0
slack-bolt==1.17.0
slack-sdk==3.20.2
